requirements.txt
numpy
opencv-python
pyyaml
dotmap
tqdm
glob
scikit-learn
scikit-image
scipy
pandas